+++ src/app/online-orders/shared/index.ts
@@ -2,4 +2,6 @@ export * from './validators/required';
 export * from './validators/scheduling';
 export * from './models/schedule-rule';
 export * from './models/transport-rule';
+export * from './models/transport-provider';
+export * from './models/transport-provider-availability';
 export * from './models/record';
