+++ src/app/online-orders/shared/models/transport-provider-availability.ts
@@ -1,0 +1,6 @@
+import { Record } from './record';
+
+export class TransportProviderAvailability extends Record<TransportProviderAvailability> {
+    day: number;
+    available: boolean;
+}
