+++ src/app/online-orders/shared/models/transport-rule.ts
@@ -21,17 +21,4 @@ export enum TransportType {
     transport_pickup
 }
 
-export class TransportProvider extends Record<TransportProvider> {
-    key: string;
-    text: string;
-    defaultAttribute: boolean;
-    sortorder?: number;
-    active: boolean;
-    availabilityRules: TransportProviderAvailability[];
-}
-
-export class TransportProviderAvailability extends Record<TransportProviderAvailability> {
-    day: number;
-    available: boolean;
-}
